feat(service): 예약 시간 외래키 참조 관련 예외 처리 로직 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,16 +4,24 @@ import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.dto.ReservationResponseDto;
 import roomescape.dto.ReservationTimeResponseDto;
 import roomescape.entity.ReservationTime;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.ReservationTimeException;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(
+        final ReservationRepository reservationRepository,
+        final ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -36,6 +44,9 @@ public class ReservationTimeService {
 
     @Transactional
     public void removeReservationTime(final long id) {
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new ReservationTimeException(ErrorCode.RESERVATION_TIME_IS_REFERENCED);
+        }
         reservationTimeRepository.deleteById(id);
     }
 }
